requirements.txt
Django==1.4.3
gunicorn==0.17.2
MySQL-python==1.2.4
django-annoying==0.7.6
django-registration==0.8
django-admin-bootstrapped==0.3.2
django-registration-defaults==0.3
django-appconf==0.5
django-pagination==1.0.7
django-compressor==1.2
django-evolution==0.6.7
django-tastypie==0.9.11
python-dateutil==1.5
mimeparse==0.1.3
requests==1.1.0
boto==2.8.0
BeautifulSoup==3.2.0